Issue #95 resolvida
Agora secretario(a) e coordenador(a) podem visualizar e escolher gerar
PDF das bancas marcadas.

+++ src/br/ufjf/tcc/pdfHandle/Ata.java
@@ -58,14 +58,17 @@ public abstract class Ata {
 				System.out.println("NAO deletado");
 		}
 		
-		for (int i = 0; i < trabMarcados.size(); i++) {
-			d = new File(ConfHandler.getConf("FILE.PATH") + "last" + idAluno
-					+ "-" + i + ".pdf");
-			if (d.delete())
-				System.out.println("deletado");
-			else
-				System.out.println("NAO deletado");
+		if(trabMarcados!=null){
+			for (int i = 0; i < trabMarcados.size(); i++) {
+				d = new File(ConfHandler.getConf("FILE.PATH") + "last" + idAluno
+						+ "-" + i + ".pdf");
+				if (d.delete())
+					System.out.println("deletado");
+				else
+					System.out.println("NAO deletado");
+			}			
 		}
+		
 
 		d = new File(ConfHandler.getConf("FILE.PATH") + "saida" + idAluno
 				+ ".pdf");
